+++ assets/javascripts/kitten/karl/pages/homepage/index.js
@@ -1,15 +1,18 @@
 import React from 'react'
 import styled from 'styled-components'
 import { Title } from '../../../components/typography/title'
+import { ArrowIcon } from '../../../components/icons/arrow-icon'
 import { pxToRem, stepToRem } from '../../../helpers/utils/typography'
 import { Container } from '../../../components/grid/container'
 import { Grid, GridCol } from '../../../components/grid/grid'
 import { ScreenConfig } from '../../../constants/screen-config'
 import COLORS from '../../../constants/colors-config'
+import TYPOGRAPHY from '../../../constants/typography-config'
 import { Video } from './video'
 import { NewsBlock } from './news/components/news-block'
 import { DialogWithPictureBorder } from './dialog-with-picture-border'
 import { FeaturedProjects } from './featured-projects'
+import { PopularProjects } from './carousel'
 import { TextWithSideImage } from './text-with-side-image'
 
 const StyledHomePage = styled.div`
@@ -61,13 +64,44 @@ const StyledHorizontalStroke = styled.div`
 `
 
 const StyledSectionTitle = styled(Title)`
-  margin: 0 0 ${pxToRem(10)};
   font-size: ${stepToRem(-1)};
 
   @media (min-width: ${pxToRem(ScreenConfig.S.min)}) {
-    margin: 0 0 ${pxToRem(15)};
     font-size: ${stepToRem(0)};
   }
+`
+
+const StyledCarouselLink = styled.a`
+  display: block;
+  text-align: right;
+  text-decoration: none;
+  ${TYPOGRAPHY.fontStyles.regular}
+  color: ${COLORS.font1};
+  transition: color .4s ease;
+
+  span {
+    font-size: ${stepToRem(-2)};
+    margin-right: ${pxToRem(7)};
+
+    @media (min-width: ${pxToRem(ScreenConfig.S.min)}) {
+      font-size: ${stepToRem(-1)};
+      margin-right: ${pxToRem(8)};
+    }
+  }
+
+  &:hover,
+  &:focus,
+  &:active {
+    color: ${COLORS.primary2}
+  }
+
+  svg {
+    vertical-align: center;
+    fill: currentColor;
+    transition: fill .4s ease;
+  }
+
+
 `
 
 const HomePage = () => (
@@ -86,7 +120,14 @@ const HomePage = () => (
     </Container>
 
     <Container as="section" className="Homepage__Section">
-      <StyledSectionTitle tag="h2" margin={false}>
+      <StyledSectionTitle
+        tag="h2"
+        className="
+          k-u-margin-top-none
+          k-u-margin-bottom-singleHalf@s-up
+          k-u-margin-bottom-single@s-down
+        "
+      >
         Notre actu
       </StyledSectionTitle>
       <NewsBlock
@@ -100,8 +141,7 @@ const HomePage = () => (
               'https://source.unsplash.com/random/300x400?kitten,1',
             imageUrlLandscape:
               'https://source.unsplash.com/random/600x300?kitten,1',
-            url:
-              '#'
+            url: '#',
           },
           {
             title: 'Cras mattis consectetur purus sit amet fermentum.',
@@ -111,8 +151,7 @@ const HomePage = () => (
               'https://source.unsplash.com/random/300x400?kitten,2',
             imageUrlLandscape:
               'https://source.unsplash.com/random/600x300?kitten,2',
-            url:
-              '#'
+            url: '#',
           },
           {
             title: 'Donec sed odio dui.',
@@ -122,8 +161,7 @@ const HomePage = () => (
               'https://source.unsplash.com/random/300x400?kitten,3',
             imageUrlLandscape:
               'https://source.unsplash.com/random/600x300?kitten,3',
-            url:
-              '#'
+            url: '#',
           },
           {
             title:
@@ -134,23 +172,48 @@ const HomePage = () => (
               'https://source.unsplash.com/random/300x400?kitten,4',
             imageUrlLandscape:
               'https://source.unsplash.com/random/600x300?kitten,4',
-            url:
-              '#'
+            url: '#',
           },
         ]}
       />
     </Container>
+
+    <section className="Homepage__Section">
+      <Container
+        className="k-u-margin-bottom-singleHalf@s-up
+          k-u-margin-bottom-single@s-down"
+      >
+        <Grid>
+          <GridCol col={8}>
+            <StyledSectionTitle tag="h2" margin={false}>
+              Nos projets populaires
+            </StyledSectionTitle>
+          </GridCol>
+          <GridCol col={4}>
+            <StyledCarouselLink href="#">
+              <span className="k-u-hidden@xs-down">Voir tous les projets</span>
+              <span className="k-u-hidden@s-up" aria-hidden>
+                Tous les projets
+              </span>
+              <ArrowIcon />
+            </StyledCarouselLink>
+          </GridCol>
+        </Grid>
+      </Container>
+
+      <PopularProjects />
+    </section>
 
     <Container as="section" className="Homepage__Section">
       <TextWithSideImage />
     </Container>
 
     <Container as="section" className="Homepage__Section">
-      <Video/>
+      <Video />
     </Container>
 
     <section className="Homepage__Section Homepage__LastSection">
-      <DialogWithPictureBorder/>
+      <DialogWithPictureBorder />
     </section>
   </StyledHomePage>
 )